requirements.txt
opencv-python~=4.10.0.84
Flask~=3.0.3
spidev~=3.6
RPi.GPIO~=0.7.1